pyproject.toml
[tool.poetry]
name = "shopping-design-system"
version = "0.1.0"
description = "An shopping design system"
authors = ["Nam Tran"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
Django = ">=4.2.13,<4.3.0"
djangorestframework = ">=3.15.2,<3.16.0"
psycopg2 = "^2.9.6"
Pillow = "^9.5.0"
django-allauth = "^0.54.0"
djangorestframework-simplejwt = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^7.3.0"
pytest-django = "^4.5.2"
black = "^23.3.0"
flake8 = "^6.1.0"
isort = "^5.11.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80
target-version = ['py38']

[tool.isort]
profile = "black"
#
#[build-system]
#requires = ["poetry-core"]
#build-backend = "poetry.core.masonry.api"
